Working chatUI & Calendar
In this version the work is all about UI, so the keyboard has been fixed in Andriod where previsoly it is was not rising the input. The calendar works well but I am concerend about the perfomance. I hope after exporting it as a standalone app it may perform better!